fix(payment): PAYPAL-2451 tests + updated after review
