+++ packages/braintree-integration/src/components/BraintreeAchPaymentForm/BraintreeAchPaymentForm.test.tsx
@@ -3,7 +3,7 @@ import {
     CheckoutService,
     createCheckoutService,
 } from '@bigcommerce/checkout-sdk';
-import { fireEvent, render, screen } from '@testing-library/react';
+import { act, fireEvent, render, screen } from '@testing-library/react';
 import { Formik } from 'formik';
 import { FormikValues } from 'formik/dist/types';
 import { noop } from 'lodash';
@@ -44,6 +44,9 @@ describe('BraintreeAchPaymentForm', () => {
             paymentForm,
             language,
             updateMandateText: jest.fn().mockReturnValue('mandate text'),
+            isInstrumentFeatureAvailable: false,
+            outstandingBalance: 10,
+            storeName: 'test',
         };
 
         jest.spyOn(checkoutService.getState().data, 'getBillingAddress').mockReturnValue({
@@ -176,5 +179,110 @@ describe('BraintreeAchPaymentForm', () => {
         expect(screen.queryByTestId('checkout-ach-form').parentElement).toHaveStyle(
             'display: none',
         );
+    });
+
+    describe('Guest Flow', () => {
+        it('should show form without Vaulting component and without ability to save instrument', async () => {
+            render(<BraintreeAchPaymentFormTest {...defaultProps} />);
+
+            await act(async () => {
+                await new Promise((resolve) => setTimeout(resolve, 0));
+            });
+
+            expect(screen.queryByText(/Account number ending in:/)).not.toBeInTheDocument();
+            expect(screen.queryByText(/Routing Number:/)).not.toBeInTheDocument();
+
+            expect(screen.getByTestId('checkout-ach-form')).toBeInTheDocument();
+            expect(
+                screen.queryByText('Save this card for future transactions'),
+            ).not.toBeInTheDocument();
+        });
+    });
+
+    describe('Vaulting Flow', () => {
+        beforeEach(() => {
+            defaultProps.isInstrumentFeatureAvailable = true;
+        });
+
+        it('should select default first instrument if there is no default instrument', async () => {
+            jest.spyOn(checkoutState.data, 'getInstruments').mockReturnValue([
+                {
+                    bigpayToken: '12341234',
+                    provider: 'braintree',
+                    accountNumber: '0001',
+                    issuer: '000100019',
+                    trustedShippingAddress: true,
+                    defaultInstrument: false,
+                    method: 'ach',
+                    type: 'bank',
+                },
+                {
+                    bigpayToken: '121341',
+                    provider: 'braintree',
+                    accountNumber: '0002',
+                    issuer: '000100010',
+                    trustedShippingAddress: true,
+                    defaultInstrument: false,
+                    method: 'ach',
+                    type: 'bank',
+                },
+            ]);
+
+            render(<BraintreeAchPaymentFormTest {...defaultProps} />);
+
+            await act(async () => {
+                await new Promise((resolve) => setTimeout(resolve, 0));
+            });
+
+            expect(screen.getByText('Account number ending in: 0001')).toBeInTheDocument();
+            expect(screen.getByText('Routing Number: 000100019')).toBeInTheDocument();
+        });
+
+        it('should show the default instrument if there is one', async () => {
+            jest.spyOn(checkoutState.data, 'getInstruments').mockReturnValue([
+                {
+                    bigpayToken: '12341234',
+                    provider: 'braintree',
+                    accountNumber: '0001',
+                    issuer: '000100019',
+                    trustedShippingAddress: true,
+                    defaultInstrument: false,
+                    method: 'ach',
+                    type: 'bank',
+                },
+                {
+                    bigpayToken: '121341',
+                    provider: 'braintree',
+                    accountNumber: '0002',
+                    issuer: '000100010',
+                    trustedShippingAddress: true,
+                    defaultInstrument: true,
+                    method: 'ach',
+                    type: 'bank',
+                },
+            ]);
+
+            render(<BraintreeAchPaymentFormTest {...defaultProps} />);
+
+            await act(async () => {
+                await new Promise((resolve) => setTimeout(resolve, 0));
+            });
+
+            expect(screen.getByText('Account number ending in: 0002')).toBeInTheDocument();
+            expect(screen.getByText('Routing Number: 000100010')).toBeInTheDocument();
+        });
+
+        it('should show form without Vaulting component', async () => {
+            jest.spyOn(checkoutState.data, 'getInstruments').mockReturnValue([]);
+
+            render(<BraintreeAchPaymentFormTest {...defaultProps} />);
+
+            await act(async () => {
+                await new Promise((resolve) => setTimeout(resolve, 0));
+            });
+
+            expect(screen.getByTestId('checkout-ach-form')).toBeInTheDocument();
+            expect(screen.getByText('Save this card for future transactions')).toBeInTheDocument();
+        });
     });
 });
